Fixed up profile styles
