+++ client/src/Pages/Profile.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import { withStyles } from '@material-ui/core/styles';
 import { useParams } from 'react-router';
-import { Button } from '@material-ui/core';
 import { gql, useQuery } from "@apollo/client";
 import { useToasts } from "react-toast-notifications";
-import MailIcon from '@material-ui/icons/Mail';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import {
+  ButtonBox,
+  BackArrow,
+  TextBox,
+  MailBox,
   ProfileCard,
 	ReturnHeader,
 	UserName,
@@ -16,43 +16,6 @@ import {
 	StyledText2,
 	StyledText3
 } from './ProfileStyles.js';
-
-const ButtonBox = withStyles({
-  label: {
-    textTransform: 'capitalize',
-  }
-})(Button);
-
-const BackArrow = withStyles({
-  root: {
-    display: 'flex',
-    color: '#2075D8',
-    justifyContent: 'center',
-    alignItems: 'center'
-  }
-})(ChevronLeftIcon);
-
-const TextBox = withStyles({
-  root: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: '25px',
-    width: '75vw',
-    height: '10vh',
-    borderRadius: '9px',
-    background: 'rgba(187, 218, 255, 0.22)'
-  }, 
-  label: {
-    textTransform: 'none',
-  }
-})(Button);
-
-const MailBox = withStyles({
-  root: {
-    color: '#2075D8'
-  }
-})(MailIcon);
 
 const Profile = () => {
